+++ CHANGELOG.md
@@ -15,6 +15,7 @@
     1. File System (Allow KMS Key Selection)
     2. Block Volume (Allow KMS Key Selection)
     3. Object Storage Bucket (Allow KMS Key Selection)
+    4. Load Balancer - Allow specification of multiple Backend Sets and Listeners.
 5. Update Discovery Code to include additional resources requires the python module to be upgraded to the latest release.
 6. Remove Object Storage from PCA flagged OKIT designs.
 7. Update Terraform Generation to create multiple files based on OCI groupings.